JCBC-1381: Properly close trace span of subdoc_mutate ops.
Motivation
----------
The single subdoc_mutate ops (which include ops like insert and upsert)
opened a span properly, but never closed them when the response arrived.

As a result, they would never show up in the tracer and so are also not
included in the slowlog threshold logging tracer.

Modifications
-------------
Similar to the other ops already in place, the regular subdoc_mutate
subdoc command now correctly closes the span, making it appear in the
threshold logging tracer if needed.

Change-Id: I49efb8998c5e8016b719823ea321ba6fd756c951
Reviewed-on: http://review.couchbase.org/112569
Reviewed-by: Graham Pople <grahampople@gmail.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>