JCBC-1594: Retry query error 4040.
It has been discovered that this error is returned by a 6.5 or later
cluster even if the client sends the prepared statement paylods (because
it does not look at them anymore due to security reasons).

If the client receives a 4040, it is expected to clear the cache entry
for this specific prepared statement and reprepare transparently.

Change-Id: I10fe886fa43e971b444b59b5f96149fce694b85c
Reviewed-on: http://review.couchbase.org/122597
Reviewed-by: Graham Pople <grahampople@gmail.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>