JCBC-1642: ClusterManager must use resolved SeedNodes
Previously the cluster manager would use the seed nodes directly
from the connection string, but this is a problem because if
DNS SRV is used it would try to connect to the dns host instead
of the resolved hosts.

This changeset updates the code so that the potentially resolved
seed hosts are used instead of the original one from the connection
string.

Note that this is still not 100% accurate since the cluster topology
might drift over time, but it fixes the issue at hand. Further
improvements can be maded alongside if neeeded in a later
changeset.

Change-Id: Ie96186d72b01c783172544433a4c456be15ced95
Reviewed-on: http://review.couchbase.org/c/couchbase-java-client/+/129906
Reviewed-by: Graham Pople <grahampople@gmail.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>